requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==24.0
pillow==10.2.0
proglog==0.1.10
PyPDF2==3.0.1
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
